requirement.txt
flask
sqlalchemy
Flask-SQLAlchemy
ConfigParser
pycrypto
python-highcharts
wtforms
https://pypi.python.org/packages/0f/7c/b316cd9779817173e93f5cebc8fb387db33cc8dc526f3db5e61f2c008d5b/Werkzeug-0.11.9.tar.gz#md5=e4dbeb6302ce74babc0d7c21fc3d8291
flask_mail
itsdangerous
datetime
flask-wtf
wtforms

possible more as located in lib